level_07_observability/requirements.txt
strands-agents[otel]>=1.0.0
strands-agents-tools>=0.1.0
boto3>=1.35.0
bedrock-agentcore>=1.0.0
aws-opentelemetry-distro
mcp>=1.0.0
requests
PyJWT
